Remove unboxing

+++ modules/asc/src/main/java/macromedia/asc/embedding/Shell.java
@@ -240,9 +240,9 @@ public class Shell
 				Object memTypeObj = getTypeMeth.invoke(memPoolObj, emptyObj);
 				Long used = (Long) getUsedMeth.invoke(memUsageObj, emptyObj);
 				if (heapField.get(typeCls) == memTypeObj) {
-					heapUsed += used.longValue();
+					heapUsed += used;
 				} else {
-					nonHeapUsed += used.longValue();
+					nonHeapUsed += used;
 				}
 				resetPeakUsageMeth.invoke(memPoolObj, emptyObj);
 			}
