+++ modules/compiler/src/main/java/flex2/tools/FontLicenseChecker.java
@@ -51,7 +51,7 @@ public class FontLicenseChecker
 			Map.Entry entry = (Map.Entry) integerArrayListEntry;
 			Integer integ = (Integer) entry.getKey();
 			ArrayList list = (ArrayList) entry.getValue();
-			FSType type = FSType.getFSType(integ.intValue());
+			FSType type = FSType.getFSType(integ);
 			String licenseType = type.description + ", fsType = '" + type.fsType + "'";
 			System.out.println("");
 			System.out.println("---------");
